raspberry-detection/requirements.txt
gradio
ultralytics
opencv-python-headless
paho-mqtt
picamera2